backend/requirements.txt
# Core Backend
django>=4.2.7
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0

# Machine Learning
torch>=2.6.0
torchvision>=0.21.0
numpy>=1.26.1
pandas>=2.1.2
scikit-learn>=1.3.2
opencv-python-headless>=4.8.1.78
albumentations>=1.3.1
matplotlib>=3.8.1
seaborn>=0.13.0
pillow>=10.1.0

# API
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart==0.0.6

# Utilities
tqdm==4.66.1
requests==2.31.0
